requirements.txt
anthropic>=0.39.0
pandas>=2.0.0
openpyxl>=3.1.0
odfpy>=1.4.1
streamlit>=1.30.0
python-dotenv>=1.0.0
plotly>=5.18.0
